Revert changes for shipping to keep this PRs focus narrower
