+++ client/extensions/woocommerce/state/data-layer/ui/woocommerce-services/index.js
@@ -5,16 +5,79 @@
  */
 
 import { isEmpty } from 'lodash';
+import { translate } from 'i18n-calypso';
 /**
  * Internal dependencies
  */
 import { WOOCOMMERCE_SERVICES_SHIPPING_ACTION_LIST_CREATE } from 'woocommerce/state/action-types';
-import { actionListStepNext, actionListClear } from 'woocommerce/state/action-list/actions';
-import { getSaveSettingsActionListSteps } from './reducer';
+import {
+	actionListStepNext,
+	actionListStepSuccess,
+	actionListStepFailure,
+	actionListClear,
+} from 'woocommerce/state/action-list/actions';
 import {
 	areLabelsEnabled,
+	getLabelSettingsFormMeta,
 	getSelectedPaymentMethodId,
 } from 'woocommerce/woocommerce-services/state/label-settings/selectors';
+import { getPackagesForm } from 'woocommerce/woocommerce-services/state/packages/selectors';
+import { submit as submitLabels } from 'woocommerce/woocommerce-services/state/label-settings/actions';
+import { submit as submitPackages } from 'woocommerce/woocommerce-services/state/packages/actions';
+import { getSelectedSiteId } from 'state/ui/selectors';
+
+const getSaveLabelSettingsActionListSteps = ( state, siteId ) => {
+	const labelFormMeta = getLabelSettingsFormMeta( state, siteId );
+	if ( ! labelFormMeta || labelFormMeta.pristine || ! labelFormMeta.can_manage_payments ) {
+		return [];
+	}
+
+	return [
+		{
+			description: translate( 'Saving label settings' ),
+			onStep: ( dispatch, actionList ) => {
+				dispatch(
+					submitLabels(
+						siteId,
+						() => dispatch( actionListStepSuccess( actionList ) ),
+						() => dispatch( actionListStepFailure( actionList ) )
+					)
+				);
+			},
+		},
+	];
+};
+
+const getSavePackagesActionListSteps = ( state, siteId ) => {
+	const packagesForm = getPackagesForm( state, siteId );
+	if ( ! packagesForm || packagesForm.pristine ) {
+		return [];
+	}
+
+	return [
+		{
+			description: translate( 'Saving label settings' ),
+			onStep: ( dispatch, actionList ) => {
+				dispatch(
+					submitPackages(
+						siteId,
+						() => dispatch( actionListStepSuccess( actionList ) ),
+						() => dispatch( actionListStepFailure( actionList ) )
+					)
+				);
+			},
+		},
+	];
+};
+
+const getSaveSettingsActionListSteps = state => {
+	const siteId = getSelectedSiteId( state );
+
+	return [
+		...getSaveLabelSettingsActionListSteps( state, siteId ),
+		...getSavePackagesActionListSteps( state, siteId ),
+	];
+};
 
 export default {
 	[ WOOCOMMERCE_SERVICES_SHIPPING_ACTION_LIST_CREATE ]: [
